Run all subtasks with HyperQueue using a dependency graph
The dependency DiGraph is generated using NetworkX, which makes it easy to look
up the dependencies of any given subtask.

Currently, this simply assumes that each task depends on the previous one for a
given species/channel. But this could be improved by actually parsing the task
line to see where the geometry optimizations occur, and using those as
dependencies for all subsequent jobs at the same geometry.